+++ backend/src/firemerge/statement/parser.py
@@ -93,16 +93,22 @@ class StatementParser:
     ) -> Iterable[StatementTransaction]:
         iban_col = self.role_cols.get(ColumnRole.IBAN)
         # prepare rows for join if doc number is present
+        join_rows: dict[object, Sequence[ValueType]] | None = None
+        duplicates: set[object] | None = None
         if join_col := self.role_cols.get(ColumnRole.DOC_NUMBER):
             assert iban_col is not None
             rows = list(rows)
-            join_rows = {
-                row[join_col.index]: row
-                for row in rows
-                if row[join_col.index] and row[iban_col.index] != self.account.iban
-            }
-        else:
-            join_rows = None
+            join_rows = {}
+            duplicates = set()
+            for scan_row in rows:
+                if (
+                    (doc_number := scan_row[join_col.index])
+                    and scan_row[iban_col.index] != self.account.iban
+                    and self._get_amount(scan_row)
+                ):
+                    if doc_number in join_rows:
+                        duplicates.add(doc_number)
+                    join_rows[doc_number] = scan_row
 
         for row, next_row in pairwise(chain(rows, [None])):
             assert row is not None  # only next_row can be None
@@ -115,7 +121,12 @@ class StatementParser:
                 and (doc_number := row[join_col.index])
                 and (join_row := join_rows.get(doc_number))
             ):
-                transaction = self._parse_row(row, join_row)
+                if duplicates and doc_number in duplicates:
+                    logger.warning("Duplicate doc number: %s", doc_number)
+                    transaction = self._parse_row(row)
+                else:
+                    logger.debug("Joining rows by doc_number=%s", doc_number)
+                    transaction = self._parse_row(row, join_row)
             else:
                 transaction = self._parse_row(row)